+++ app/global.scss
@@ -4,6 +4,7 @@
   --sub-black: #333;
   --alpha-black: rgba(37, 37, 37, 0.5);
 }
+
 * {
   box-sizing: border-box;
   padding: 0;
@@ -11,6 +12,9 @@
 }
 html,
 body {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
   max-width: 100vw;
   overflow-x: hidden;
 }
@@ -21,6 +25,17 @@ body {
     Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+}
+main {
+  flex: 1;
+}
+li {
+  list-style-type: none;
+  a {
+    &:hover {
+      opacity: 0.8;
+    }
+  }
 }
 a {
   color: inherit;
